fix: copier 재시도와 atomic 파일 처리 개선

+++ internal/syncer/copier/copier.go
@@ -11,8 +11,11 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/avast/retry-go"
+
 	"github.com/schollz/progressbar/v3"
 
+	"amuz.es/src/spi-ca/fast-volume-syncer/internal/args"
 	"amuz.es/src/spi-ca/fast-volume-syncer/internal/util"
 
 	"amuz.es/src/spi-ca/fast-volume-syncer/internal/returns"
@@ -56,110 +59,21 @@ type Copier struct {
 	SourceRoot      string
 	DestinationRoot string
 	Umask           os.FileMode
+	Retry           args.RetryArgs
 	opIdx           uint64
 }
 
-func (t *Copier) copyNewFile(opIdx uint64, srcPath, dstPath string, mode os.FileMode) (int64, error) {
-	dstDir := filepath.Dir(dstPath)
-	src, err := os.OpenFile(srcPath, os.O_RDONLY, 0o644)
-	if err == nil {
-		// do nothing
-	} else if os.IsNotExist(err) {
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   errors.Join(ErrCopierSrcNotExist, err),
-		}
-	} else {
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   fmt.Errorf("failed to open source file: %w; %w", ErrCopierCopyFailed, err),
-		}
+func (t *Copier) Execute(ctx context.Context, fileList []returns.Fileinfo) error {
+	if t.Retry.Attempts <= 0 {
+		return t.execute(ctx, fileList)
 	}
 
-	tm, err := times.StatFile(src)
-	if err == nil {
-		// do nothing
-	} else if os.IsNotExist(err) {
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   errors.Join(ErrCopierSrcNotExist, err),
-		}
-	} else {
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   fmt.Errorf("failed to get the source fileinfo :%w; %w", ErrCopierCopyFailed, err),
-		}
-	}
-	defer src.Close()
-
-	tmp, err := os.CreateTemp(dstDir, fmt.Sprintf(".tmp-%x-%d", int64(os.Getpid())^time.Now().Unix(), opIdx))
-	if err != nil {
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   fmt.Errorf("failed to create a tempfile :%w; %w", ErrCopierCopyFailed, err),
-		}
-	}
-
-	tmpPath := tmp.Name()
-	copied, err := io.Copy(tmp, src)
-	if err == nil {
-		// do nothing
-	} else if os.IsNotExist(err) {
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   errors.Join(ErrCopierSrcNotExist, err),
-		}
-	} else {
-		_ = tmp.Close()
-		_ = os.Remove(tmpPath)
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   errors.Join(ErrCopierCopyFailed, err),
-		}
-	}
-
-	err = os.Rename(tmpPath, dstPath)
-	if err != nil {
-		_ = tmp.Close()
-		_ = os.Remove(tmpPath)
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   fmt.Errorf("failed to rename a file :%w; %w", ErrCopierCopyFailed, err),
-		}
-	}
-	_ = tmp.Close()
-
-	err = os.Chmod(dstPath, mode)
-	if err != nil {
-		_ = os.Remove(dstPath)
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   fmt.Errorf("failed to change a filemode :%w; %w", ErrCopierCopyFailed, err),
-		}
-	}
-
-	err = os.Chtimes(dstPath, tm.AccessTime(), tm.ModTime())
-	if err != nil {
-		_ = os.Remove(dstPath)
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   fmt.Errorf("failed to update times :%w; %w", ErrCopierCopyFailed, err),
-		}
-	}
-	return copied, nil
+	retryOptionArgs := t.Retry.Assemble(ctx)
+	retryFunc := func() error { return t.execute(ctx, fileList) }
+	return retry.Do(retryFunc, retryOptionArgs...)
 }
 
-func (t *Copier) copyExistingFile(opIdx uint64, srcPath, dstPath string, mode os.FileMode) (int64, error) {
+func (t *Copier) copyFile(parentCtx context.Context, opIdx uint64, srcPath, dstPath string, mode os.FileMode, removeDst bool) (int64, error) {
 	dstDir := filepath.Dir(dstPath)
 	src, err := os.OpenFile(srcPath, os.O_RDONLY, 0o644)
 	if err == nil {
@@ -206,39 +120,59 @@ func (t *Copier) copyExistingFile(opIdx uint64, srcPath, dstPath string, mode os
 	}
 
 	tmpPath := tmp.Name()
-	copied, err := io.Copy(tmp, src)
-	if err == nil {
+	var copied int64
+	ctx, causeFunc := context.WithCancelCause(context.Background())
+	go func(copiedPtr *int64) {
+		copiedBytes, copyErr := io.Copy(tmp, src)
 		// do nothing
-	} else if os.IsNotExist(err) {
-		return 0, &copierError{
+		*copiedPtr = copiedBytes
+		causeFunc(copyErr)
+	}(&copied)
+	select {
+	case <-parentCtx.Done():
+		err = &copierError{
 			srcPath: srcPath,
 			dstPath: dstPath,
-			cause:   errors.Join(ErrCopierSrcNotExist, err),
+			cause:   parentCtx.Err(),
 		}
-	} else {
-		_ = tmp.Close()
-		_ = os.Remove(tmpPath)
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   errors.Join(ErrCopierCopyFailed, err),
+	case <-ctx.Done():
+		cancelReason := context.Cause(ctx)
+		if errors.Is(cancelReason, context.Canceled) {
+			break
+		} else if os.IsNotExist(cancelReason) {
+			err = &copierError{
+				srcPath: srcPath,
+				dstPath: dstPath,
+				cause:   errors.Join(ErrCopierSrcNotExist, err),
+			}
+		} else {
+			err = &copierError{
+				srcPath: srcPath,
+				dstPath: dstPath,
+				cause:   errors.Join(ErrCopierCopyFailed, err),
+			}
 		}
 	}
-
-	err = os.Remove(dstPath)
+	_ = tmp.Close()
 	if err != nil {
-		_ = tmp.Close()
 		_ = os.Remove(tmpPath)
-		return 0, &copierError{
-			srcPath: srcPath,
-			dstPath: dstPath,
-			cause:   fmt.Errorf("failed to remove the destination :%w; %w", ErrCopierCopyFailed, err),
+		return 0, err
+	}
+
+	if removeDst {
+		err = os.RemoveAll(dstPath)
+		if err == nil {
+			_ = os.Remove(tmpPath)
+			return 0, &copierError{
+				srcPath: srcPath,
+				dstPath: dstPath,
+				cause:   fmt.Errorf("failed to remove the destination :%w; %w", ErrCopierCopyFailed, err),
+			}
 		}
 	}
 
 	err = os.Rename(tmpPath, dstPath)
 	if err != nil {
-		_ = tmp.Close()
 		_ = os.Remove(tmpPath)
 		return 0, &copierError{
 			srcPath: srcPath,
@@ -246,7 +180,6 @@ func (t *Copier) copyExistingFile(opIdx uint64, srcPath, dstPath string, mode os
 			cause:   fmt.Errorf("failed to rename a file :%w; %w", ErrCopierCopyFailed, err),
 		}
 	}
-	_ = tmp.Close()
 
 	err = os.Chmod(dstPath, mode)
 	if err != nil {
@@ -267,7 +200,6 @@ func (t *Copier) copyExistingFile(opIdx uint64, srcPath, dstPath string, mode os
 			cause:   fmt.Errorf("failed to update times :%w; %w", ErrCopierCopyFailed, err),
 		}
 	}
-
 	return copied, nil
 }
 
@@ -307,18 +239,20 @@ func (t *Copier) compareFile(opIdx uint64, srcPath string, dstPath string, srcSi
 		}
 	}
 
-	if srcTm.ModTime().Before(dstTm.ModTime()) {
-		return COMPARE_DST_IS_NEWER, nil
-	}
-
 	if srcSize != destMode.Size() {
 		return COMPARE_DIFFER, nil
-	} else {
+	}
+
+	if offset := srcTm.ModTime().Sub(dstTm.ModTime()); offset == 0 {
 		return COMPARE_EQUAL, nil
+	} else if offset < 0 {
+		return COMPARE_DST_IS_NEWER, nil
+	} else {
+		return COMPARE_DIFFER, nil
 	}
 }
 
-func (t *Copier) copyRegularFile(opIdx uint64, srcPath string, dstPath string, srcSize int64, dstMode os.FileMode) (int64, error) {
+func (t *Copier) copyRegularFile(ctx context.Context, opIdx uint64, srcPath string, dstPath string, srcSize int64, dstMode os.FileMode) (int64, error) {
 	differ, err := t.compareFile(opIdx, srcPath, dstPath, srcSize)
 	if err != nil {
 		return 0, err
@@ -338,9 +272,9 @@ func (t *Copier) copyRegularFile(opIdx uint64, srcPath string, dstPath string, s
 			cause:   errors.Join(ErrCopierSrcNotExist, err),
 		}
 	case COMPARE_DIFFER:
-		return t.copyExistingFile(opIdx, srcPath, dstPath, dstMode)
+		return t.copyFile(ctx, opIdx, srcPath, dstPath, dstMode, true)
 	case COMPARE_DST_NOT_EXIST:
-		return t.copyNewFile(opIdx, srcPath, dstPath, dstMode)
+		return t.copyFile(ctx, opIdx, srcPath, dstPath, dstMode, false)
 	case COMPARE_DST_IS_NEWER:
 		util.ErrLog.Printf("[Copier op:%d]destination(%s) is newer than source(%s)", opIdx, dstPath, dstMode)
 		return 0, nil
@@ -361,7 +295,8 @@ func (t *Copier) processDirectory(opIdx uint64, srcPath string, dstPath string, 
 	if err == nil {
 		if destExists = destMode.IsDir(); !destExists {
 			// 대상 path가 directory mode가 아닌 경우 대상을 날린다.
-			if err = os.RemoveAll(dstPath); err != nil {
+			err = os.RemoveAll(dstPath)
+			if err == nil {
 				return &copierError{
 					srcPath: srcPath,
 					dstPath: dstPath,
@@ -424,7 +359,8 @@ func (t *Copier) processSymbolicLink(opIdx uint64, srcPath string, dstPath strin
 		}
 
 		// 대상 path가 symlink mode가 아닌 경우 대상을 날린다.
-		if err = os.RemoveAll(dstPath); err != nil {
+		err = os.RemoveAll(dstPath)
+		if err != nil {
 			return &copierError{
 				srcPath: srcPath,
 				dstPath: dstPath,
@@ -444,7 +380,8 @@ func (t *Copier) processSymbolicLink(opIdx uint64, srcPath string, dstPath strin
 	if destMode, err := os.Lstat(dirPath); err == nil {
 		if !destMode.IsDir() {
 			// 대상 path가 directory mode가 아닌 경우 대상을 날린다.
-			if err = os.RemoveAll(dirPath); err != nil {
+			err = os.RemoveAll(dirPath)
+			if err != nil {
 				return &copierError{
 					srcPath: srcPath,
 					dstPath: dstPath,
@@ -477,7 +414,7 @@ func (t *Copier) processSymbolicLink(opIdx uint64, srcPath string, dstPath strin
 	return nil
 }
 
-func (t *Copier) routeFileByTypes(opIdx uint64, srcInfo returns.Fileinfo) (int64, error) {
+func (t *Copier) routeFileByTypes(ctx context.Context, opIdx uint64, srcInfo returns.Fileinfo) (int64, error) {
 	srcMode := srcInfo.Mode
 	dstMode := srcInfo.Mode.Perm() | t.Umask
 
@@ -489,7 +426,7 @@ func (t *Copier) routeFileByTypes(opIdx uint64, srcInfo returns.Fileinfo) (int64
 	} else if srcMode.Type()&fs.ModeSymlink != 0 {
 		return 0, t.processSymbolicLink(opIdx, srcPath, dstPath, srcInfo.SymlinkPath)
 	} else if srcMode.IsRegular() {
-		return t.copyRegularFile(opIdx, srcPath, dstPath, srcInfo.Size, dstMode)
+		return t.copyRegularFile(ctx, opIdx, srcPath, dstPath, srcInfo.Size, dstMode)
 	} else {
 		return 0, &copierError{
 			srcPath: srcPath,
@@ -499,7 +436,7 @@ func (t *Copier) routeFileByTypes(opIdx uint64, srcInfo returns.Fileinfo) (int64
 	}
 }
 
-func (t *Copier) Execute(ctx context.Context, fileList []returns.Fileinfo) error {
+func (t *Copier) execute(ctx context.Context, fileList []returns.Fileinfo) error {
 	opIdx := atomic.AddUint64(&t.opIdx, 1)
 	filenameSet := make(map[string]int)
 	for idx, info := range fileList {
@@ -529,19 +466,19 @@ func (t *Copier) Execute(ctx context.Context, fileList []returns.Fileinfo) error
 
 forLoop:
 	for _, entry := range fileList {
-		copied, err := t.routeFileByTypes(opIdx, entry)
+		copied, err := t.routeFileByTypes(ctx, opIdx, entry)
 		_ = bar.Add(1)
 		res.appendFilename(entry.Path)
 		res.sentBytes += copied
 		if err == nil {
 			res.sent++
+		} else if errors.Is(err, context.Canceled) {
 		} else if errors.Is(err, ErrCopierUptodate) {
 			res.uptodate++
 		} else if errors.Is(err, ErrCopierSrcNotExist) {
-			util.ErrLog.Printf("[Copier op:%d]%v", opIdx, err)
+			res.disappeared++
 		} else if errors.Is(err, ErrCopierSkipped) {
 			res.skipped++
-			util.ErrLog.Printf("[Copier op:%d]%v", opIdx, err)
 		} else {
 			res.errs = append(res.errs, err)
 		}
